Tyres promoted to main, Firmware to Technical
