+++ app/src/main/java/lu/fisch/canze/fragments/ExperimentalFragment.java
@@ -185,18 +185,6 @@ public class ExperimentalFragment extends Fragment {
             }
         });
 
-        button = (Button) view.findViewById(R.id.buttonTyres);
-        button.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if(!MainActivity.isSafe()) return;
-                if(MainActivity.device==null) {MainActivity.toast("You first need to adjust the settings ..."); return;}
-                MainActivity.getInstance().leaveBluetoothOn=true;
-                Intent intent = new Intent(MainActivity.getInstance(), TyresActivity.class);
-                ExperimentalFragment.this.startActivityForResult(intent, MainActivity.LEAVE_BLUETOOTH_ON);
-            }
-        });
-
         return view;
     }
