+++ app/src/main/res/layout/fragment_technical.xml
@@ -90,6 +90,39 @@
 
             </TableRow>
 
+            <TableRow
+                android:layout_width="match_parent"
+                android:layout_height="match_parent">
+
+
+                <Button
+                    android:layout_width="0dp"
+                    android:layout_weight="1"
+                    android:layout_height="fill_parent"
+                    android:layout_marginTop="4dp"
+                    android:layout_marginLeft="4dp"
+                    android:layout_marginRight="4dp"
+                    android:layout_marginBottom="4dp"
+                    android:text="Firmware"
+                    android:id="@+id/buttonFirmware"
+                    android:drawableLeft="@mipmap/button_firmware"
+                    android:layout_column="0" />
+                <Button
+                    android:id="@+id/button_2"
+                    android:layout_width="0dp"
+                    android:layout_weight="1"
+                    android:layout_height="fill_parent"
+                    android:layout_marginTop="8dp"
+                    android:layout_marginLeft="8dp"
+                    android:layout_marginRight="8dp"
+                    android:layout_marginBottom="8dp"
+                    android:text="-"
+                    android:layout_column="1" />
+
+            </TableRow>
+
+
+
         </TableLayout>
 
     </ScrollView>
